test(tests): pin the measured priority parity and correct the latch fix's rationale
CI measured what the probe test asked: on the 3-core runner a test body, a
Task { }, a Task.detached { } and a gateHoldingTask { } all run at medium
(rawValue 21). The rationale the ShutdownLatch(executor:) fix carried — that a
detached task queues behind higher-priority test tasks — is refuted.

What remains is FIFO hop latency. With every task at one priority the pool's
queue is FIFO among the ~5,000 tasks a pass keeps runnable, and one suspension
hop costs the pass's own per-test latency (p50 65.6 s, p99 92.1 s on a green
fast pass 1). The latch handshake is three or four hops deep, so 0 of 8 callers
back after 90 s is hop count times latency, not priority ordering. Executor
injection still helps, and for the right reason: it removes those hops from the
shared queue rather than reordering them within it.

- Turn the probe into a pin of the measured fact: TaskPriorityParityTests
  asserts all four readings are equal, keeps the continuation-based
  non-escalating apparatus and the .fastPassBounded guard, and its control now
  discriminates — an explicitly elevated task must read higher, so an apparatus
  reporting a constant cannot make parity green for the wrong reason.
- Rewrite the Tests/CLAUDE.md passage under "Thread-blocking gates run off the
  cooperative pool" around hop count, pointing at the pin by name. Surrounding
  guidance (inject the executor where the test owns the callee, bound at
  TestDeadlines.saturatedPass where it does not) is unchanged.
- Correct the same claim in ServerShutdownLatchTests' comments; code untouched.
